Move `ForumPost` and `ForumComment` to the forum folder
