+++ packages/ui/src/common/types/index.ts
@@ -1,6 +1,5 @@
 export * from './Activity'
 export * from './Block'
-export * from './Forum'
 
 export type Address = string
 export type EmptyObject = Record<string, never>
