+++ packages/ui/src/forum/components/ForumComment/ForumComment.tsx
@@ -9,8 +9,8 @@ import { MarkdownPreview } from '@/common/components/MarkdownPreview'
 import { Badge, TextInlineSmall } from '@/common/components/typography'
 import { Colors, Transitions } from '@/common/constants'
 import { relativeTime } from '@/common/model/relativeTime'
-import { ForumPost } from '@/common/types'
 import { spacing } from '@/common/utils/styles'
+import { ForumPost } from '@/forum/types'
 import { MemberInfo } from '@/memberships/components'
 
 import { BlockDate } from './BlockDate'
