+++ packages/ui/src/forum/components/ForumComment/index.ts
@@ -1,0 +1,1 @@
+export * from './ForumComment'
